Inline select / update sql statements
For model items stored in the AIM sqlite database, we test first for
the existence of a row in the db. This was getting out of sync with the
call to either update or insert because I had an extra transaction
around it. I’ve removed the transaction so that it happens in the same
transaction as the select.