Fix for #96, #97
Reworked import screen for projects and documents. Users can now only
import one item at a time. Documents are listed with the familiar arrow
on the side; clicking one will show the “please wait” UI, with the
status as the file is imported. Also fixed the callback for chapters,
so that each chapter is immediately available after import.