Move project model to sql
This is a possible fix for issue #112, but I won’t know until there is
another code change. The issue in #112 _might_ be due to some of the
model data being held in localStorage, which I’ve now removed.

This fix constitutes a major code change, and should be tested with a
variety of data to make sure we haven’t broken anything.