Fix #143
Export (for all but XML) loops through each chapter, exporting only if
there is at least some adaptation done on that chapter. If the export
hits the last adapted source phrase ID, it will stop exporting the
target text for that chapter. This should cover a sequential
adaptation, but might export some extra punctuation for an adaptation
that is being done out of chapter order.