Fix #194, fix #141
Output files are now in a subdirectory of the documents, shared,
external root or data directory (depending on the platform) labeled
`AIM_Exports_<rfc>` where `<rfc>` is the RFC 5646 language code for the
target language. This should differentiate the output texts from other
documents. I’ll be updating the wiki with the directory info.