Fix #264
Looks like the “no files” problem on Android >5 is due to new runtime
permissions that need to be requested. Currently the only plugin I see
that supports this is cordova.plugins.diagnostic — it will query the
device for permissions and ask if necessary. AIM now does this at app
startup.