Work on #301
Support import of clipboard text -
If there is something on the clipboard, an extra entry "Clipboard Text" appears on the list of possible files to import. If the user clicks on it, the clipboard text is pasted / imported in as a new "file" with the title "Text-#" (where # is a unique number).