Fix #341
Add "preview" button to adaptation toolbar, only visible if "show preview button" is selected in Settings>General Settings. This button toggles between a read-only target text display and the normal adaptation/editing display. For the smallest phone screens, this button replaces the help walkthrough button (<350px width); which is not really all that useful for the more experienced user.