Fix #394
This was a nastier problem than expected; the current code was working around an internal issue with marionettejs removing the page region when the model changed. The fix here is to add the region when the user displays a project settings view, and remove the region when the user clicks OK or Cancel.