Work on #328
Import into AIM seems to be working when AIM is open. Had to change reader to read as arraybuffer and then convert to UTF-8 (larger files were failing due to a bug in the plain text reader). TODO: AIM currently launches when a file is sent to it, but it doesn't do the import -- need to investigate further.