Fix #538
Using a file chooser instead of scanning the local drives for file import. Usability improved 100-fold, BUT this required a lot of refactoring. Will probably be dealing with lurking bugs for a bit before this milestone can be released (and will need to be released in lock-step with the project file issue #539).