Fix #545
Last couple fixes on the Android side: the open/save dialogs spawn in a new process, throwing AIM in the background. Had to ignore the Resume() event in these cases (removed it completely in import, ignored it conditionally in export) so the import/export wizards weren't reset. Looks like things are working on both platforms, so I'm checking in the fix. Will test further.